decouple loghttp from logql by moving validation to logql package
