+++ pkg/loghttp/series.go
@@ -37,13 +37,6 @@ func ParseSeriesQuery(r *http.Request) (*logproto.SeriesRequest, error) {
 		}
 	}
 
-	// ensure matchers are valid before fanning out to ingesters/store as well as returning valuable parsing errors
-	// instead of 500s
-	_, err = Match(deduped)
-	if err != nil {
-		return nil, err
-	}
-
 	return &logproto.SeriesRequest{
 		Start:  start,
 		End:    end,
